NoTrailingWhitespaceInCommentFixer - move to PSR2 level
